Вынесение метода POST
