+++ areacomment.js
@@ -1,0 +1,7 @@
+export function sanitizeHtml({htmlString}) {
+    return htmlString
+        .replaceAll("&", "&amp;")
+        .replaceAll("<", "&lt;")
+        .replaceAll(">", "&gt;")
+        .replaceAll('"', "&quot;");
+};
